Moved ui files around a bit.
