+++ ui/qt/new_order.ui
@@ -100,6 +100,11 @@
    </item>
    <item row="6" column="0">
     <widget class="QLabel" name="lblWhatUTXO">
+      <property name="font">
+        <font>
+        <underline>true</underline>
+        </font>
+      </property>
      <property name="cursor">
       <cursorShape>WhatsThisCursor</cursorShape>
      </property>
